Treat tilde and hyphen equivalently in semver
OCaml's version number (defined in Sys.ocaml_version) is not precisely
compatible with semantic versioning's notion of pre-release and build.

Since 4.11.0, it has been permitted for the additional-information to be
preceded with a tilde or a plus.

Relax the interpretation of semver slightly and treat hyphen and tilde
as equivalent. This means for 4.12.0 pre-releases and onwards, the
alpha1, etc. portion is placed in the _PRELEASE variable instead of
_BUILD.