Cppo_types, cppo_parser, cppo_eval: simplify the type of [`Ident].
In the definition of [`Ident] in the type [node],
the third parameter used to have type [actuals option]:
it was either [None]
or [Some args] where [args] was a nonempty list of actuals.

It is simpler to let this parameter have type [actuals].
There is no need for an option.