Refactoring: in [expand_ordinary_ident], remove [g.require_location := false].
The previous call to [expand_list ... (text ...)] always leaves
[g.require_location] set to [false]. (Proof: [text ...] always
ends with a non-space [`Text] item, and processing such an item
sets [g.require_location] to [false].)

So, this change should not have any impact, and it makes the code
cleaner. Ideally, [expand_ordinary_ident] should be a one-line
function. We are almost there.